+++ frontend/src/components/proof-templates/RestrictionsEdit.vue
